Fixes #7333,BZ1129635 - Deleting a resource and creation a new one wont lead to conflicts

+++ app/lib/actions/katello/content_view_environment/destroy.rb
@@ -17,26 +17,37 @@ module Actions
 
         def plan(cv_env, options = {})
           skip_cp_update = options.fetch(:skip_candlepin_update, false)
+          skip_repo_destroy = options.fetch(:skip_repo_destroy, false)
+          organization_destroy = options.fetch(:organization_destroy, false)
           content_view = cv_env.content_view
           environment = cv_env.environment
-          content_view.check_remove_from_environment!(environment)
+          content_view.check_remove_from_environment!(environment) unless organization_destroy
 
           sequence do
             concurrence do
-              content_view.repos(environment).each do |repo|
-                # no need to update the content view environment since it's
-                # getting destroyed so skip_environment_update
-                plan_action(Repository::Destroy, repo, skip_environment_update: true)
+              unless skip_repo_destroy
+                content_view.repos(environment).each do |repo|
+                  # no need to update the content view environment since it's
+                  # getting destroyed so skip_environment_update
+                  plan_action(Repository::Destroy, repo, skip_environment_update: true)
+                end
               end
 
               if puppet_env = content_view.puppet_env(environment)
                 plan_action(Actions::Foreman::Environment::Destroy, puppet_env.puppet_environment) if puppet_env.puppet_environment
-                plan_action(ContentViewPuppetEnvironment::Destroy, puppet_env)
+                plan_action(ContentViewPuppetEnvironment::Destroy, puppet_env) unless organization_destroy
               end
             end
             plan_action(Candlepin::Environment::Destroy, cp_id: cv_env.cp_id) unless skip_cp_update
+            plan_self(:id => cv_env.id)
+          end
+        end
 
-            cv_env.reload
+        def finalize
+          cv_env = ::Katello::ContentViewEnvironment.find_by_id(input[:id])
+          if cv_env.nil?
+            output[:response] = "Content view with ID #{input[:id]} is (probably) already deleted"
+          else
             cv_env.destroy!
           end
         end
